first step of changing code architecture. followed the MVVM nodel: M - Model - the data structure. V- View - the GUI , has no code behind VM- View Model - the logic of the application.
The VM has tight coupling to the Model and a reference to the View. the VM should be the main point of the application.
The view should not have reference to the VM, only to the properties of the Model.
Currently the view still contains the middleman - the main VM.

have to decouple it in next commits.