Remade the Tags database. Now the database contains a table o ftags as well as rules. the rules table has a connection to the tags table. each rule is connected to one spesific tag. Was able to read both the Tags and the Rules in the SQLiteconnection.
Remade the Tagger using the refection and the rules i set up.
Made an "EveryTHingElse" rule that makes a function that sends true everytime. this function must always be last.

For finishing, i made a function that checks if the Tag.db is empty - if so it popuplates it with default rules and tags.

Next Steps:
1) Must correct the visuals now - have to somehow add the ComputerTime tagModel and timer.
2) must make adding rules to exisitng rules available. AND KEEPING THE EVERYTHING ELSE RULE LAST
3) must make adding tags window with rules - can't add a tag without a rule.